dogs rare
With 190 recognized dog breeds, it can be difficult to learn about them all. While the United States is filled with Labrador Retrievers and German Shepherd Dogs, two of the most popular dog breeds in the country, there are many other breeds that go relatively unnoticed. If it weren’t for their enthusiasts, we might not even see as many as we do. But just because they’re less popular, doesn’t mean these dogs have nothing to offer. In fact, it’s quite the opposite. The rarest dog breeds in the United States are actually some of the most exciting!

English Foxhound